pyproject.toml
[tool.poetry]
name = "visualization"
version = "0.1.0"
description = ""
authors = ["Jefferson Qiu <jeffersonqiu1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.33.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.14"
openai = "^1.16.2"
plotly = "^5.20.0"
importlib = "^1.0.4"
langchain-experimental = "^0.0.56"
tabulate = "^0.9.0"
langchain-community = "^0.0.31"
statsmodels = "^0.14.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
